requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
azure-ai-documentintelligence==1.0.0b4
azure-ai-language==1.0.0
azure-ai-translation-text==1.0.0
azure-cognitiveservices-search-newssearch==2.0.0
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
openai==1.3.0
python-dotenv==1.0.0
lxml==4.9.3
xmltodict==0.13.0
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

